Changed names of the files for Admin LTE menu renders - task #809
